refactor render module
