+++ include/render.h
@@ -6,15 +6,17 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/23 00:09:35 by reasuke           #+#    #+#             */
-/*   Updated: 2024/06/23 00:24:34 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/06/23 21:03:31 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef RENDER_H
 # define RENDER_H
 
+# include <stdbool.h>
+
 # include "ctx.h"
 
-void	render(t_ctx *ctx);
+void	render(t_ctx *ctx, bool is_rereder);
 
 #endif
